HDDS-1357. ozone s3 shell command has confusing subcommands. Contributed by Elek, Marton.
